+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Support for Python 3.3 has been dropped, it reached its end of life.
 
 ### Harmonization
 - Renamed `JSON` to `JSONDict` and added a new type `JSON`. `JSONDict` saves data internally as JSON, but acts like a dictionary. `JSON` accepts any valid JSON.
+- added destination.urlpath and source.urlpath to harmonization.
 
 #### Parsers
 - changed feednames in `bots.parsers.shadowserver`. Please refer to it's README for the exact changes.